+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## 1.0 branch
 ### 1.0.2-beta3
 * added `TagEntity`, which now represents the tags;
+* tags are now a collection (so `AbstractEntity::getTagsByName()` returns now a
+    `Collection` instance);
 * added `AbstractEntity::hasTag()` method;
 * added `ClassEntity::getType()` method;
 * added `ClassEntity::createFromName()` static method;